requirements.txt
py-cord==2.4.1
NewFunctionsPYC==1.3.2
discord-py-slash-command==4.2.1